Code restructure and file creation based on model output
