+++ .gitignore
@@ -25,3 +25,4 @@ replay_pid*
 
 # Kotlin Gradle plugin data, see https://kotlinlang.org/docs/whatsnew20.html#new-directory-for-kotlin-data-in-gradle-projects
 .kotlin/
+/Web_Gen/src/main/resources/generated
