+++ Web_Gen/src/main/kotlin/app/web_gen/code_generation/CodeSnippetRepository.kt
@@ -1,4 +1,4 @@
-package app.web_gen
+package app.web_gen.code_generation
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
