+++ Web_Gen/src/main/kotlin/app/web_gen/exceprions/FileAlreadyExistsException.kt
@@ -1,0 +1,3 @@
+package app.web_gen.exceprions
+
+data class FileAlreadyExistsException(var filename:String): Exception("File $filename already Exists")
